refactored via restructuring of package engine to create subpackages expander and evaluator
