+++ src/battlesnake.theserverproject.com/insert_clever_name/snake/engine/engine.go
@@ -1,72 +1,19 @@
 package engine
 
 import (
+	"battlesnake.theserverproject.com/insert_clever_name/snake/engine/evaluator"
+	"battlesnake.theserverproject.com/insert_clever_name/snake/engine/expander"
+	"battlesnake.theserverproject.com/insert_clever_name/snake/game"
 	"context"
 	"fmt"
 	"math"
-	"math/rand"
 	"sync"
 	"time"
-
-	"battlesnake.theserverproject.com/insert_clever_name/snake/game"
 )
 
-type node struct {
-	Game     game.Game
-	Children []*node
-	Expanded bool
-	Move     Move // this is the move type that generated this board from the previous board
-	Terminal bool
-}
-
-func expandTree(ctx context.Context, n *node, depth int, maximizingPlayer bool, wg *sync.WaitGroup) {
-	defer wg.Done()
+func alphabeta(n *expander.Node, depth int, alpha float64, beta float64, maximizingPlayer bool) float64 {
 	if depth == 0 || n.Terminal {
-		return
-	}
-	if !n.Expanded {
-		if maximizingPlayer {
-			resetTurn(n.Game)
-		}
-		for childGame := range nextGameStates(ctx, n.Game, maximizingPlayer) {
-			select {
-			case <-ctx.Done():
-				return
-			default:
-				terminalNode := false
-				// if we are dead or we are the only snake left, path is terminated
-				if _, ok := childGame.ValueSnakeMap[game.ME]; !ok || len(childGame.ValueSnakeMap) == 1 {
-					terminalNode = true
-				}
-				childRef := &node{
-					Game: childGame,
-					Children: nil,
-					Expanded: false,
-					Terminal: terminalNode,
-				}
-				n.Children = append(n.Children, childRef)
-				wg.Add(1)
-				go expandTree(ctx, childRef, depth-1, !maximizingPlayer, wg)
-			}
-		}
-		n.Expanded = true
-		n.Game.Board = nil
-		n.Game.ValueSnakeMap = nil
-	} else {
-		for _, child := range n.Children {
-			wg.Add(1)
-			go expandTree(ctx, child, depth-1, !maximizingPlayer, wg)
-		}
-	}
-}
-
-func evaluate(g game.Game) float64 {
-	return float64(rand.Intn(100))
-}
-
-func alphabeta(n *node, depth int, alpha float64, beta float64, maximizingPlayer bool) float64 {
-	if depth == 0 || n.Terminal {
-		return evaluate(n.Game)
+		return evaluator.Evaluate(n.Game)
 	}
 	if maximizingPlayer {
 		value := math.Inf(-1) // negative infinity
@@ -91,15 +38,15 @@ func alphabeta(n *node, depth int, alpha float64, beta float64, maximizingPlayer
 	}
 }
 
-func ComputeMove(g game.Game, deadline time.Duration) Move {
+func ComputeMove(g game.Game, deadline time.Duration) expander.Move {
 	ctx, cancel := context.WithTimeout(context.Background(), deadline*time.Millisecond) // process the move for x ms, leaving (500 - x) ms for the network
 	defer cancel()
-	root := node{
+	root := expander.Node{
 		Game:     g,
 		Children: nil,
 		Expanded: false,
 	}
-	latestMove := UP // default move is some arbitrary direction for now
+	latestMove := expander.UP // default move is some arbitrary direction for now
 	depth := 2
 	evaluated := make(chan int, 1)
 	expanded := make(chan int, 1)
@@ -114,7 +61,7 @@ func ComputeMove(g game.Game, deadline time.Duration) Move {
 			go func() {
 				start := time.Now().UnixNano() / int64(time.Millisecond)
 				wg.Add(1)
-				expandTree(ctx, &root, depth, true, &wg)
+				expander.Expand(ctx, &root, depth, true, &wg)
 				wg.Wait()
 				end := time.Now().UnixNano() / int64(time.Millisecond)
 				fmt.Printf("Expansion took %d milliseconds for depth %d\n", end - start, depth)
